Add Preamble and Postamble to SLO and SegmentedIterable
This patch updates the SLO middleware and SegmentedIterable to add
support for user specified preamble and postamble data for each
segment. If specified the preamble/postamble data are expected to
be base64 encoded data to be added before/after an object segment
within an SLO. To accommodate the potential extra preamble and
postamble data we increase the default SLO maximum manifest size
from 2MiB to 8MiB.

This patch is a prerequisite for a future patch enabling the
download of large objects as tarballs. The TLO patch will be added
as a dependent patch later.

Change-Id: Ib8dc216a84d370e6da7d6b819af79582b671d699